Adds an esm bundle (#80)
This changes the rollup config so that it can build an esm bundle. This
excludes the canvas module which gets shimmed to a noop.

You can run with:

```
npx rollup -c rollup/es.config.j
```

I didn't add a build script or publish hook as I wasn't sure how you
felt about those, figured you'd adjust that to your liking anyways.

Adds a test that this works in a web worker. I've tested in Deno as well
and it works there too.

Cheers!

Closes #79

Co-authored-by: Matthew Phillips <matthew@phillipsoft.com>