Event bubbling keep target (#89)
* when bubbling keep original target

for event bubbling we need to keep the original target of the event. For that we read it from the event argument and if it's defined it's kept for next dispatch. If not defined, its value is "this".

+ added define helper from ungap/event-target

* events: improve bubbling

- remove Event and EventTarget Node extend to ensure consistency.
This is to avoid problems with lack of bubbling support from Node,
while also allowing support for correct "target" value.

- added simple "Event.composedPath()" and also "dispatchEvent that doesn't rely on recursive call (because it was a bit hacky to implement the correct target that way);

- unit tests now check composedPath and target / currentTarget.

* events: testing eventPhase

* events: dispatchEvent return only based on defaultPrevented

* events: dispatchEvent return only based on defaultPrevented

* events: event-target update + dictionary

* events: removing ungap dependency

* events: line break at the end of dictionary file

* events: revert and back to create(null)

* events: removing traces of dictionary

* events: types update

* events: update after npm install

* events: revert line break at the end of package.json

* events: revert package.json changes

* events: remove again ungap event-target

I reverted package.json changes because I had weird changes in that file when doing "npm i", but after updating node to 16.5 from 14.x everything is fine again.
Also the build seems to fail but after some investigation it doesn't look directly linked to PR's changes.

* events: update after package.json change

* events: using more Map and WeakMap return

* events: revert space and assert trace

* events: removing useless "set"